x-trove_web/backend/requirements.txt
Flask==2.3.3
xrpl-py==1.10.0
python-dotenv==1.0.0
flask-cors